add a new IAM example to support S3 topic
