+++ javav2/example_code/iam/src/main/java/com/example/iam/scenario/IAMScenario.java
@@ -1,0 +1,416 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package com.example.iam.scenario;
+
+// snippet-start:[iam.java2.scenario.import]
+import software.amazon.awssdk.auth.credentials.AwsBasicCredentials;
+import software.amazon.awssdk.auth.credentials.AwsSessionCredentials;
+import software.amazon.awssdk.auth.credentials.StaticCredentialsProvider;
+import software.amazon.awssdk.core.waiters.WaiterResponse;
+import software.amazon.awssdk.regions.Region;
+import software.amazon.awssdk.services.iam.IamClient;
+import software.amazon.awssdk.services.iam.model.AccessKey;
+import software.amazon.awssdk.services.iam.model.AttachRolePolicyRequest;
+import software.amazon.awssdk.services.iam.model.AttachedPolicy;
+import software.amazon.awssdk.services.iam.model.CreateAccessKeyRequest;
+import software.amazon.awssdk.services.iam.model.CreateAccessKeyResponse;
+import software.amazon.awssdk.services.iam.model.CreatePolicyRequest;
+import software.amazon.awssdk.services.iam.model.CreatePolicyResponse;
+import software.amazon.awssdk.services.iam.model.CreateRoleRequest;
+import software.amazon.awssdk.services.iam.model.CreateRoleResponse;
+import software.amazon.awssdk.services.iam.model.CreateUserRequest;
+import software.amazon.awssdk.services.iam.model.CreateUserResponse;
+import software.amazon.awssdk.services.iam.model.DeleteAccessKeyRequest;
+import software.amazon.awssdk.services.iam.model.DeletePolicyRequest;
+import software.amazon.awssdk.services.iam.model.DeleteRoleRequest;
+import software.amazon.awssdk.services.iam.model.DeleteUserRequest;
+import software.amazon.awssdk.services.iam.model.DetachRolePolicyRequest;
+import software.amazon.awssdk.services.iam.model.GetPolicyRequest;
+import software.amazon.awssdk.services.iam.model.GetPolicyResponse;
+import software.amazon.awssdk.services.iam.model.GetUserRequest;
+import software.amazon.awssdk.services.iam.model.GetUserResponse;
+import software.amazon.awssdk.services.iam.model.IamException;
+import software.amazon.awssdk.services.iam.model.ListAttachedRolePoliciesRequest;
+import software.amazon.awssdk.services.iam.model.ListAttachedRolePoliciesResponse;
+import software.amazon.awssdk.services.iam.model.User;
+import software.amazon.awssdk.services.iam.waiters.IamWaiter;
+import software.amazon.awssdk.services.s3.S3Client;
+import software.amazon.awssdk.services.s3.model.ListObjectsRequest;
+import software.amazon.awssdk.services.s3.model.ListObjectsResponse;
+import software.amazon.awssdk.services.s3.model.S3Object;
+import software.amazon.awssdk.services.sts.StsClient;
+import software.amazon.awssdk.services.sts.model.AssumeRoleRequest;
+import software.amazon.awssdk.services.sts.model.AssumeRoleResponse;
+import software.amazon.awssdk.services.sts.model.Credentials;
+import software.amazon.awssdk.services.sts.model.StsException;
+import java.util.List;
+import java.util.concurrent.TimeUnit;
+// snippet-end:[iam.java2.scenario.import]
+
+// snippet-start:[iam.java2.scenario.main]
+/*
+  To run this Java V2 code example, set up your development environment, including your credentials.
+
+  For information, see this documentation topic:
+
+  https://docs.aws.amazon.com/sdk-for-java/latest/developer-guide/get-started.html
+
+  This example performs these operations:
+
+  1. Creates a user that has no permissions.
+  2. Creates a role and policy that grants Amazon S3 permissions.
+  3. Creates a role.
+  4. Grants the user permissions.
+  5. Gets temporary credentials by assuming the role.  Creates an Amazon S3 Service client object with the temporary credentials.
+  6. Deletes the resources.
+ */
+
+public class IAMScenario {
+    public static final String DASHES = new String(new char[80]).replace("\0", "-");
+    public static final String PolicyDocument = "{" +
+            "  \"Version\": \"2012-10-17\"," +
+            "  \"Statement\": [" +
+            "    {" +
+            "        \"Effect\": \"Allow\"," +
+            "        \"Action\": [" +
+            "            \"s3:*\"" +
+            "       ]," +
+            "       \"Resource\": \"*\"" +
+            "    }" +
+            "   ]" +
+            "}";
+
+    public static String userArn;
+
+    public static void main(String[] args) throws Exception {
+
+        final String usage = """
+
+                Usage:
+                    <username> <policyName> <roleName> <roleSessionName> <bucketName>\s
+
+                Where:
+                    username - The name of the IAM user to create.\s
+                    policyName - The name of the policy to create.\s
+                    roleName - The name of the role to create.\s
+                    roleSessionName - The name of the session required for the assumeRole operation.\s
+                    bucketName - The name of the Amazon S3 bucket from which objects are read.\s
+                """;
+
+        if (args.length != 5) {
+            System.out.println(usage);
+            System.exit(1);
+        }
+
+        String userName = args[0];
+        String policyName = args[1];
+        String roleName = args[2];
+        String roleSessionName = args[3];
+        String bucketName = args[4];
+
+        Region region = Region.AWS_GLOBAL;
+        IamClient iam = IamClient.builder()
+                .region(region)
+                .build();
+
+        System.out.println(DASHES);
+        System.out.println("Welcome to the AWS IAM example scenario.");
+        System.out.println(DASHES);
+
+        System.out.println(DASHES);
+        System.out.println(" 1. Create the IAM user.");
+        User createUser = createIAMUser(iam, userName);
+
+        System.out.println(DASHES);
+        userArn = createUser.arn();
+
+        AccessKey myKey = createIAMAccessKey(iam, userName);
+        String accessKey = myKey.accessKeyId();
+        String secretKey = myKey.secretAccessKey();
+        String assumeRolePolicyDocument = "{" +
+                "\"Version\": \"2012-10-17\"," +
+                "\"Statement\": [{" +
+                "\"Effect\": \"Allow\"," +
+                "\"Principal\": {" +
+                "	\"AWS\": \"" + userArn + "\"" +
+                "}," +
+                "\"Action\": \"sts:AssumeRole\"" +
+                "}]" +
+                "}";
+
+        System.out.println(assumeRolePolicyDocument);
+        System.out.println(userName + " was successfully created.");
+        System.out.println(DASHES);
+        System.out.println("2. Creates a policy.");
+        String polArn = createIAMPolicy(iam, policyName);
+        System.out.println("The policy " + polArn + " was successfully created.");
+        System.out.println(DASHES);
+
+        System.out.println(DASHES);
+        System.out.println("3. Creates a role.");
+        TimeUnit.SECONDS.sleep(30);
+        String roleArn = createIAMRole(iam, roleName, assumeRolePolicyDocument);
+        System.out.println(roleArn + " was successfully created.");
+        System.out.println(DASHES);
+
+        System.out.println(DASHES);
+        System.out.println("4. Grants the user permissions.");
+        attachIAMRolePolicy(iam, roleName, polArn);
+        System.out.println(DASHES);
+
+        System.out.println(DASHES);
+        System.out.println("*** Wait for 30 secs so the resource is available");
+        TimeUnit.SECONDS.sleep(30);
+        System.out.println("5. Gets temporary credentials by assuming the role.");
+        System.out.println("Perform an Amazon S3 Service operation using the temporary credentials.");
+        assumeRole(roleArn, roleSessionName, bucketName, accessKey, secretKey);
+        System.out.println(DASHES);
+
+        System.out.println(DASHES);
+        System.out.println("6 Getting ready to delete the AWS resources");
+        deleteKey(iam, userName, accessKey);
+        deleteRole(iam, roleName, polArn);
+        deleteIAMUser(iam, userName);
+        System.out.println(DASHES);
+
+        System.out.println(DASHES);
+        System.out.println("This IAM Scenario has successfully completed");
+        System.out.println(DASHES);
+    }
+
+    public static AccessKey createIAMAccessKey(IamClient iam, String user) {
+        try {
+            CreateAccessKeyRequest request = CreateAccessKeyRequest.builder()
+                    .userName(user)
+                    .build();
+
+            CreateAccessKeyResponse response = iam.createAccessKey(request);
+            return response.accessKey();
+
+        } catch (IamException e) {
+            System.err.println(e.awsErrorDetails().errorMessage());
+            System.exit(1);
+        }
+        return null;
+    }
+
+    public static User createIAMUser(IamClient iam, String username) {
+        try {
+            // Create an IamWaiter object
+            IamWaiter iamWaiter = iam.waiter();
+            CreateUserRequest request = CreateUserRequest.builder()
+                    .userName(username)
+                    .build();
+
+            // Wait until the user is created.
+            CreateUserResponse response = iam.createUser(request);
+            GetUserRequest userRequest = GetUserRequest.builder()
+                    .userName(response.user().userName())
+                    .build();
+
+            WaiterResponse<GetUserResponse> waitUntilUserExists = iamWaiter.waitUntilUserExists(userRequest);
+            waitUntilUserExists.matched().response().ifPresent(System.out::println);
+            return response.user();
+
+        } catch (IamException e) {
+            System.err.println(e.awsErrorDetails().errorMessage());
+            System.exit(1);
+        }
+        return null;
+    }
+
+    public static String createIAMRole(IamClient iam, String rolename, String json) {
+
+        try {
+            CreateRoleRequest request = CreateRoleRequest.builder()
+                    .roleName(rolename)
+                    .assumeRolePolicyDocument(json)
+                    .description("Created using the AWS SDK for Java")
+                    .build();
+
+            CreateRoleResponse response = iam.createRole(request);
+            System.out.println("The ARN of the role is " + response.role().arn());
+            return response.role().arn();
+
+        } catch (IamException e) {
+            System.err.println(e.awsErrorDetails().errorMessage());
+            System.exit(1);
+        }
+        return "";
+    }
+
+    public static String createIAMPolicy(IamClient iam, String policyName) {
+        try {
+            // Create an IamWaiter object.
+            IamWaiter iamWaiter = iam.waiter();
+            CreatePolicyRequest request = CreatePolicyRequest.builder()
+                    .policyName(policyName)
+                    .policyDocument(PolicyDocument).build();
+
+            CreatePolicyResponse response = iam.createPolicy(request);
+            GetPolicyRequest polRequest = GetPolicyRequest.builder()
+                    .policyArn(response.policy().arn())
+                    .build();
+
+            WaiterResponse<GetPolicyResponse> waitUntilPolicyExists = iamWaiter.waitUntilPolicyExists(polRequest);
+            waitUntilPolicyExists.matched().response().ifPresent(System.out::println);
+            return response.policy().arn();
+
+        } catch (IamException e) {
+            System.err.println(e.awsErrorDetails().errorMessage());
+            System.exit(1);
+        }
+        return "";
+    }
+
+    public static void attachIAMRolePolicy(IamClient iam, String roleName, String policyArn) {
+        try {
+            ListAttachedRolePoliciesRequest request = ListAttachedRolePoliciesRequest.builder()
+                    .roleName(roleName)
+                    .build();
+
+            ListAttachedRolePoliciesResponse response = iam.listAttachedRolePolicies(request);
+            List<AttachedPolicy> attachedPolicies = response.attachedPolicies();
+            String polArn;
+            for (AttachedPolicy policy : attachedPolicies) {
+                polArn = policy.policyArn();
+                if (polArn.compareTo(policyArn) == 0) {
+                    System.out.println(roleName + " policy is already attached to this role.");
+                    return;
+                }
+            }
+
+            AttachRolePolicyRequest attachRequest = AttachRolePolicyRequest.builder()
+                    .roleName(roleName)
+                    .policyArn(policyArn)
+                    .build();
+
+            iam.attachRolePolicy(attachRequest);
+            System.out.println("Successfully attached policy " + policyArn + " to role " + roleName);
+
+        } catch (IamException e) {
+            System.err.println(e.awsErrorDetails().errorMessage());
+            System.exit(1);
+        }
+    }
+
+    // Invoke an Amazon S3 operation using the Assumed Role.
+    // snippet-start:[iam.java2.scenario.assumeRole]
+    public static void assumeRole(String roleArn, String roleSessionName, String bucketName, String keyVal,
+            String keySecret) {
+
+        // Use the creds of the new IAM user that was created in this code example.
+        AwsBasicCredentials credentials = AwsBasicCredentials.create(keyVal, keySecret);
+        StsClient stsClient = StsClient.builder()
+                .region(Region.US_EAST_1)
+                .credentialsProvider(StaticCredentialsProvider.create(credentials))
+                .build();
+
+        try {
+            AssumeRoleRequest roleRequest = AssumeRoleRequest.builder()
+                    .roleArn(roleArn)
+                    .roleSessionName(roleSessionName)
+                    .build();
+
+            AssumeRoleResponse roleResponse = stsClient.assumeRole(roleRequest);
+            Credentials myCreds = roleResponse.credentials();
+            String key = myCreds.accessKeyId();
+            String secKey = myCreds.secretAccessKey();
+            String secToken = myCreds.sessionToken();
+
+            // List all objects in an Amazon S3 bucket using the temp creds retrieved by
+            // invoking assumeRole.
+            Region region = Region.US_EAST_1;
+            S3Client s3 = S3Client.builder()
+                    .credentialsProvider(
+                            StaticCredentialsProvider.create(AwsSessionCredentials.create(key, secKey, secToken)))
+                    .region(region)
+                    .build();
+
+            System.out.println("Created a S3Client using temp credentials.");
+            System.out.println("Listing objects in " + bucketName);
+            ListObjectsRequest listObjects = ListObjectsRequest.builder()
+                    .bucket(bucketName)
+                    .build();
+
+            ListObjectsResponse res = s3.listObjects(listObjects);
+            List<S3Object> objects = res.contents();
+            for (S3Object myValue : objects) {
+                System.out.println("The name of the key is " + myValue.key());
+                System.out.println("The owner is " + myValue.owner());
+            }
+
+        } catch (StsException e) {
+            System.err.println(e.getMessage());
+            System.exit(1);
+        }
+    }
+
+    // snippet-end:[iam.java2.scenario.assumeRole]
+    public static void deleteRole(IamClient iam, String roleName, String polArn) {
+
+        try {
+            // First the policy needs to be detached.
+            DetachRolePolicyRequest rolePolicyRequest = DetachRolePolicyRequest.builder()
+                    .policyArn(polArn)
+                    .roleName(roleName)
+                    .build();
+
+            iam.detachRolePolicy(rolePolicyRequest);
+
+            // Delete the policy.
+            DeletePolicyRequest request = DeletePolicyRequest.builder()
+                    .policyArn(polArn)
+                    .build();
+
+            iam.deletePolicy(request);
+            System.out.println("*** Successfully deleted " + polArn);
+
+            // Delete the role.
+            DeleteRoleRequest roleRequest = DeleteRoleRequest.builder()
+                    .roleName(roleName)
+                    .build();
+
+            iam.deleteRole(roleRequest);
+            System.out.println("*** Successfully deleted " + roleName);
+
+        } catch (IamException e) {
+            System.err.println(e.awsErrorDetails().errorMessage());
+            System.exit(1);
+        }
+    }
+
+    public static void deleteKey(IamClient iam, String username, String accessKey) {
+        try {
+            DeleteAccessKeyRequest request = DeleteAccessKeyRequest.builder()
+                    .accessKeyId(accessKey)
+                    .userName(username)
+                    .build();
+
+            iam.deleteAccessKey(request);
+            System.out.println("Successfully deleted access key " + accessKey +
+                    " from user " + username);
+
+        } catch (IamException e) {
+            System.err.println(e.awsErrorDetails().errorMessage());
+            System.exit(1);
+        }
+    }
+
+    public static void deleteIAMUser(IamClient iam, String userName) {
+        try {
+            DeleteUserRequest request = DeleteUserRequest.builder()
+                    .userName(userName)
+                    .build();
+
+            iam.deleteUser(request);
+            System.out.println("*** Successfully deleted " + userName);
+
+        } catch (IamException e) {
+            System.err.println(e.awsErrorDetails().errorMessage());
+            System.exit(1);
+        }
+    }
+}
+// snippet-end:[iam.java2.scenario.main]
