+++ javav2/example_code/iam/src/test/java/IAMServiceTest.java
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import com.example.iam.*;
+import com.example.iam.scenario.IAMScenario;
 import com.google.gson.Gson;
 import org.junit.jupiter.api.*;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
